Rename io_fleur_inpgen to fleur_ingpen to avoid repetition in import
